+++ CHANGELOG.md
@@ -11,10 +11,14 @@
 
 ### New features
 
+### Bugfixes
+
+## 0.7.3
+
+### New features
+
 - Added `\Wizaplace\User\User::getTitle`
 - Added (optional) `title` parameter to `\Wizaplace\User\UserService::updateUser`
-
-### Bugfixes
 
 ## 0.7.2